# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Complaint Routing Prediction Contract",
    "version": "0.1.0",
    "description": "Contract for ClaimVox local inference. The FastAPI service is implemented for local execution; it is not deployed or a public production service."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/v1/predictions": {
      "post": {
        "operationId": "createPrediction",
        "summary": "Classify one complaint narrative",
        "x-implementation-status": "local-implemented",
        "x-data-handling": {
          "request_body_logging": "prohibited",
          "narrative_persistence": "none-by-default",
          "model_inputs": [
            "narrative"
          ]
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prediction created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Read non-sensitive service health",
        "x-implementation-status": "local-implemented",
        "responses": {
          "200": {
            "description": "Service health",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/status": {
      "get": {
        "operationId": "getOperationalStatus",
        "summary": "Read aggregate operational connection status",
        "x-implementation-status": "integrated",
        "responses": {
          "200": {
            "description": "Non-sensitive service, model and database status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "createDemoSession",
        "summary": "Create an environment-configured local demo session",
        "x-implementation-status": "integrated-demo-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Short-lived demo bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credentials rejected without disclosing which field failed"
          }
        }
      }
    },
    "/v1/feedback": {
      "post": {
        "operationId": "createLocalFeedback",
        "summary": "Record privacy-minimised local human feedback",
        "x-implementation-status": "local-implemented",
        "x-data-handling": {
          "narrative_persistence": "prohibited",
          "identity_persistence": "prohibited",
          "individual_record_export": "prohibited",
          "operation_scope": "local-only"
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback recorded locally without returning an individual record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/FeedbackRejected"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/feedback/summary": {
      "get": {
        "operationId": "getLocalFeedbackSummary",
        "summary": "Read aggregate-only local feedback summary",
        "x-implementation-status": "local-implemented",
        "x-data-handling": {
          "individual_record_export": "prohibited",
          "operation_scope": "local-only"
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregate feedback counters after expired-record purge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackSummaryResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Environment-configured single-user demo token; not a production identity system."
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "writeOnly": true
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "role",
          "name"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "const": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "service_health",
          "service_version",
          "model_version",
          "database_connected",
          "database_type"
        ],
        "properties": {
          "service_health": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ]
          },
          "service_version": {
            "type": "string"
          },
          "model_version": {
            "type": "string"
          },
          "database_connected": {
            "type": "boolean"
          },
          "database_type": {
            "type": "string",
            "enum": [
              "postgresql",
              "none"
            ]
          }
        }
      },
      "PredictionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "narrative"
        ],
        "properties": {
          "narrative": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000,
            "pattern": "\\S",
            "description": "Complaint narrative. Clients must reject whitespace-only values and warn against unnecessary personal data."
          },
          "client_request_id": {
            "type": "string",
            "maxLength": 100,
            "description": "Optional technical correlation identifier; never a model feature."
          }
        }
      },
      "CanonicalClass": {
        "type": "string",
        "enum": [
          "Checking or savings account",
          "Credit card",
          "Credit reporting or other personal consumer reports",
          "Debt collection",
          "Debt or credit management",
          "Money transfer, virtual currency, or money service",
          "Mortgage",
          "Payday loan, title loan, personal loan, or advance loan",
          "Prepaid card",
          "Student loan",
          "Vehicle loan or lease"
        ]
      },
      "NullableConfidence": {
        "oneOf": [
          {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          {
            "type": "null"
          }
        ],
        "description": "Calibrated model confidence when available; otherwise null."
      },
      "Alternative": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "class_label",
          "confidence"
        ],
        "properties": {
          "class_label": {
            "$ref": "#/components/schemas/CanonicalClass"
          },
          "confidence": {
            "$ref": "#/components/schemas/NullableConfidence"
          }
        }
      },
      "PredictionResponse": {
        "type": "object",
        "additionalProperties": false,
        "allOf": [
          {
            "if": {
              "properties": {
                "confidence": {
                  "type": "null"
                }
              },
              "required": [
                "confidence"
              ]
            },
            "then": {
              "properties": {
                "review_required": {
                  "const": true
                }
              }
            }
          }
        ],
        "required": [
          "prediction_id",
          "predicted_class",
          "alternatives",
          "confidence",
          "review_required",
          "review_reasons",
          "model_version",
          "taxonomy_version",
          "created_at",
          "warnings"
        ],
        "properties": {
          "prediction_id": {
            "type": "string",
            "format": "uuid"
          },
          "predicted_class": {
            "$ref": "#/components/schemas/CanonicalClass"
          },
          "alternatives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Alternative"
            }
          },
          "confidence": {
            "$ref": "#/components/schemas/NullableConfidence"
          },
          "review_required": {
            "type": "boolean"
          },
          "review_reasons": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "low_confidence",
                "confidence_unavailable",
                "out_of_domain",
                "language_policy",
                "service_policy"
              ]
            }
          },
          "model_version": {
            "type": "string",
            "minLength": 1
          },
          "taxonomy_version": {
            "type": "string",
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "The response never echoes the complaint narrative or recommends an operational queue."
      },
      "FeedbackCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "prediction_id",
          "model_version",
          "taxonomy_version",
          "suggested_class",
          "decision",
          "purpose"
        ],
        "properties": {
          "prediction_id": {
            "type": "string",
            "format": "uuid"
          },
          "model_version": {
            "type": "string",
            "minLength": 1
          },
          "taxonomy_version": {
            "type": "string",
            "minLength": 1
          },
          "suggested_class": {
            "$ref": "#/components/schemas/CanonicalClass"
          },
          "reviewed_class": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CanonicalClass"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision": {
            "type": "string",
            "enum": [
              "confirmed",
              "corrected",
              "not_actionable"
            ]
          },
          "purpose": {
            "type": "string",
            "enum": [
              "human_review_quality_assurance",
              "future_retraining_candidate"
            ]
          }
        },
        "description": "Approved feedback metadata only. The request never includes narrative, identity, free text, audio, transcription, or probabilities."
      },
      "FeedbackAcceptedResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "recorded"
            ]
          }
        },
        "description": "Confirmation without a feedback identifier or individual record."
      },
      "FeedbackSummaryItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model_version",
          "suggested_class",
          "decision",
          "count"
        ],
        "properties": {
          "model_version": {
            "type": "string",
            "minLength": 1
          },
          "suggested_class": {
            "$ref": "#/components/schemas/CanonicalClass"
          },
          "decision": {
            "type": "string",
            "enum": [
              "confirmed",
              "corrected",
              "not_actionable"
            ]
          },
          "count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "FeedbackSummaryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackSummaryItem"
            }
          }
        },
        "description": "Aggregate-only feedback counts; individual records and identifiers are never returned."
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error_code",
          "message",
          "request_id"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "description": "Safe client-facing error without request content or internal details."
      },
      "HealthResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "service_version"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ]
          },
          "service_version": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Narrative rejected by an active validation policy",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "FeedbackRejected": {
        "description": "Feedback request rejected without echoing submitted content",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Request frequency exceeded the active policy",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Prediction service or approved model unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}